requirements.txt
Pillow>=8.1.0
gpiozero
apscheduler
pandas
